--- src/components/index.js
@@ -1,4 +1,0 @@
-export { default as Navbar } from "./Navbar/Navbar";
-export { default as Products } from "./Products/Products";
-export { default as Cart } from "./Cart/Cart";
-export { default as Checkout } from "./CheckoutForm/Checkout/Checkout";
