+++ src/pages/Shop/CheckoutForm/AddressForm.jsx
@@ -35,7 +35,7 @@ const AddressForm = ({ test }) => {
           </Grid>
           <br />
           <div style={{ display: "flex", justifyContent: "space-between" }}>
-            <Button component={Link} variant="outlined" to="/cart">
+            <Button component={Link} variant="outlined" to="/shop/cart">
               Back to Cart
             </Button>
             <Button type="submit" variant="contained" color="primary">
