+++ src/pages/index.js
@@ -1,0 +1,2 @@
+export { default as Shop } from "./Shop";
+export { default as Home } from "./Home";
